backend/marine_drive/requirements.txt
# Django and REST Framework
Django==4.2.24
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
Pillow==10.0.1  # For ImageField support

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# API filtering and pagination
django-filter==23.3

# Development and testing
pytest==7.4.3
pytest-django==4.5.2

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.8  # PostgreSQL adapter

# ML and data processing (optional)
numpy==1.24.3
opencv-python==4.8.1.78

# Environment management
python-decouple==3.8

# API documentation
drf-spectacular==0.26.5